Add support for macro aliases
Implement support for multi-level macro aliases, useful for the kind of
thing aliases are: shortcuts, added level of indirection here and there
for eg choosing between alternatives without copying and so on.

The new %alias macro primitive can be used to define an alias and
get the alias definition. The dual functionality is necessary because
all other references to the alias get routed to the aliased macro. With
the exception of undefine which undefines the alias itself.

Add new rpmPushMacroAlias() API for defining aliases programmatically.
The o(pts) argument is unused currently and intended for potential future
expansion to allow passing default arguments to an aliased parametric macro.