Avoid having to cast every single xrealloc() and _free() in c++
We can't infer the type on malloc() or calloc() but realloc() and
free() are passed a pointer from which we can grab the return type
and have the compiler do the work of tedious mandatory typecasting
in c++ in the early stages.

typeof() is a compiler extension but one that all the major compilers
support, and it's just a temporary crutch anyhow.