Convert man pages to specialist tooling: scdoc
Pandoc do a passable job of generating man pages out of Markdown, but
pandoc itself is a horrible build dependency for the mundane task of
generating man pages. We've avoided this build dependency by shipping
rendered man-pages in our source releases. That was not such a big deal
in the autotools ecosystem where source releases are *built*, but
that's really an oxymoron. We want to move to a model where our source
releases are really nothing more than what git-archive outputs, and
something that anybody can verify anywhere at all.

To preserve man page availability, switch over to a tool that we can
trivially add as a hard build requirement. scdoc is a trivially
buildable piece of < 1000 lines of C code and depends on nothing that
rpm doesn't otherwise depend. Being a specialized tool for man pages,
I think it also does a better job at it than pandoc. Should make CI
faster too because there are fewer dependencies to install.

This is quite a chainsaw change, the main purpose is just to make
our man pages buildable with scdoc. Making it pretty and unifying all
the wildly different styles is a job for another day.

Fixes: #3563