Add an opt-in 4.x tag compatibility option as a crutch for API users
We're renaming a bunch of more-or-less obscure tag names in 6.0 because
we have to. While this is ABI compatible, it breaks the API so somebody
somewhere might run into a situation where their software no longer
builds. Add a first-aid crutch to let such API users compile while
warning them that this is not compatible with v6 packages.

Add a small test program to test the accesses.