Add a backwards compatibility map to rpmTagGetValue()
This doesn't seem or feel right for a multitude of reasons. We perhaps
need to provide this kind of mapping for binary compatibility, but for
newly compiled programs and rpm itself it seems we should redirect these
to tag extensions that return meaningful data for both v4 and v6
packages.